finalizo lectura y creacion de datos
